Simplify package structure
This commit collapses the "crs" package into the "header" package, and
re-collapses the "schema" package into the "props" package.

It was somewhat nice to be able to write `crs.New` and `schema.New`,
but this creates too many Go packages and still doesn't create a fully
clean structure, as we still had `schema.ColumnFromFlat`. If we're
going to have names like `schema.ColumnFromFlat` as opposed to just
`column.FromFlat`, we might as well use that naming consistently
everywhere rather than mixing like `props.FromFlat` and `crs.FromFlat`
with `schema.ColumnFromFlat`.